+++ lib/glimmer-dsl-tk.rb
@@ -24,7 +24,7 @@ $LOAD_PATH.unshift(File.expand_path('..', __FILE__))
 # External requires
 require 'glimmer'
 require 'logging'
-# require 'puts_debuggerer'
+require 'puts_debuggerer' if ENV['pd'].to_s.downcase == 'true'
 require 'super_module'
 require 'tk'