+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.0.2
+
+- Notebook/frame support
+- Hello, Tab! sample
+
 ## 0.0.1
 
 - Initial Glimmer DSL for TK implementation to support Hello, World!